Update d2.md
Added a couple "User Stories" that show our first priority of providing the player with an engaging yet challenging experience, and showing inheritance with a pipe ( | ) to extend the user story of the player to the Host and Client and expand to priority 2 that the Host should be able to host a game that their friends can connect to; similarly the Client should be able to connect to a Hosting friends game. This all so they can both be players and share an engaging and challenging experience.